Fix thinko coming from 9ccfeb93
pg_basebackup.c relies on the compression level to not be 0 to decide if
compression should be used, but 9ccfeb93 missed the fact that the
default compression (Z_DEFAULT_COMPRESSION) is -1, which would be used
if specifying --gzip without --compress.

While on it, add some coverage for --gzip, as this is rather easy to
miss.

Reported-by: Christoph Berg
Discussion: https://postgr.es/m/YdhRDMLjabtXOnhY@msg.df7cb.de